### 2023-02-17 #5166 @editorialbot
[PRE REVIEW]: `repytah`: An Open-Source Python Package That Builds Aligned Hierarchies for Sequential Data Streams
**Submitting author:** @kmkinnaird (<a href="http://orcid.org/0000-0002-0435-8996">Katherine M. Kinnaird</a>)
**Repository:** https://github.com/smith-tinkerlab/repytah
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
