### 2023-02-20 #5166 @fabian-s
sorry @danielskatz  this is too far from my expertise -- I don't code in Python and know nothing about streaming data.
