### 2023-02-20 #5166 @kmkinnaird
Thanks all for getting this started! I looked through the list of people who have volunteered to review for JOSS and would like to offer the below list as some possible suggestions for reviewing. The list is ordered starting from the bottom of the shared google sheet.

- rocsg
- benjaminbolling
- edeno
- ranzhengcode
- yidan-yang
- imciner2
- ziyiyin97
- lukaszjablonski
- klb2
- max-little
- epspebble
- cherry6666
- jorgedch
- itismeghasyam
- Ykarpate
