### 2023-02-28 #5166 @jbytecode
👋👋👋 Dear @edeno and Dear @rocsg    👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**repytah: An Open-Source Python Package That Builds Aligned Hierarchies for Sequential Data Streams**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5166).

Thank you in advance!
