### 2023-03-05 #5166 @ranzhengcode
Yes! It's my honor to review this manuscript!

从 Windows 版邮件<https://go.microsoft.com/fwlink/?LinkId=550986>发送

发件人: Mehmet Hakan ***@***.***>
发送时间: 2023年3月5日 18:01
收件人: ***@***.***>
抄送: Zheng ***@***.***>; ***@***.***>
主题: Re: [openjournals/joss-reviews] [PRE REVIEW]: `repytah`: An Open-Source Python Package That Builds Aligned Hierarchies for Sequential Data Streams (Issue #5166)

👋👋👋 Dear @ranzhengcode<https://github.com/ranzhengcode> 👋👋👋

Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "repytah: An Open-Source Python Package That Builds Aligned Hierarchies for Sequential Data Streams". You can find more information at the top of this Github issue (#5166<https://github.com/openjournals/joss-reviews/issues/5166>).

Thank you in advance!
