### 2023-03-05 #5166 @jbytecode
@kmkinnaird  - Thank you for submitting your paper to JOSS.

@Rocsg, @ranzhengcode - Thank you all for accepting our invitation.

The review will start in a new GitHub issue. I will introduce the requirements and instructions there.

See you there!
