### 2023-03-06 smith-tinkerlab/repytah#426 @Rocsg
First review / Rocsg
https://github.com/openjournals/joss-reviews/issues/5213

Hello @jbytecode  @ranzhengcode @kmkinnaird

I thank Katherine and her team for submitting their software and providing it open-source. The documentation is quite extensive and the code is easy to read. I began to fulfill my reviewer checklist. I have some concerns to address:

Main text :
* There is a claim of universality, stating that repytah can process any data stream, while examples are run on a single music file (a score). While I guess the claim could be valid, authors could modulate the claim (could be used… or something like that), or provide evidence with another dataset. In addition, this could be minor, but the author state the tool process *streams*, and a user could take you to the word, and provide *actual* streams, i.e. data with no start and no end. I guess it is not suitable as repytah only processes finite matrices to manage a finite-size dissimilarity matrix. But as the term “stream” was used in the first paper from Kinnaird, I guess this data naming is accepted in the field.
Also, the example file (in repytah/data) are CSV without mentioning what they are. Are they scores? Spectrum extracts? It is unclear. Please introduce a README in the data dir. I would like to understand what I processed precisely. This point could feed the first part of the question.

* L1: I am not good at English, and I guess Kinnaird master her language, but the word “That” in the title seems weird. It sounds like we found a package in the wilderness, plugged it in, and suddenly discovered that it behaves in that way. This point is a bit about “tastes and colors”, so it’s optional.

* L4: I was confused with the extensive list of authors, but after an exhaustive search, it seems that they (quite) all contributed to the repo. This is nice, I admire the proficiency of the PI to make all them work together that way on a single object. However, I did not find out what Lillian Yushu Kiang was responsible for.

* L18 End of summary “and the code has been updated to leverage efficiencies in Python”, maybe better worded I guess, or gain a few words to get an insight of the efficiencies leveraged.

* L32 In the statement of need, maybe two or three more words to describe what is the “fingerprint task” and the “cover song task”, because here are your main “customers”.

* L33: the author should check the ref (Savard et al. 2020), which is not well spelled in the manuscript.

* L35 In the statement of need you state that the original code is MATLAB, thus implies licensing issue. But there are alternatives (octave). Did you try it out? Sometimes there are performance issues, or some functionalities of MATLAB are not integrated in octave. You should state this point, while the interest in having it available in python is still high.

* L36 and L44 Speaking of a “tradition” of building MIR in Python, is maybe abusive as it is a 6 years sequence. Maybe we can speak of a trend, or an appeal from the community? Python growing interest?

* L40 : the author should check the ref (Bittner et al. 2019), which is not well spelled in the manuscript.

* L45 “ for a given sequential data stream”, can you provide the dimensionality of the input space? I work on imaging, but I guess that repytah is not suitable for 2D+t, isn’t it ? And 3D+t ?

* L49 : [...] by cross-referencing [...]. I don’t get it. What do the authors did? Did they run some comparison of the output of the Matlab package and the output of the python package, compute some comparison scores, or check exact matching? Are these data provided ? If not, it should be stated.

* L57: I don’t get the description of the module transform, you should make it clear.

* L69 : what are “Chroma features” ? If repytah can be used for any data stream, you can understand that some of your potential users won’t understand this point. Can you add a few words about it? This point is linked to the request above for the data README.md

* L72: missing “funded by” at the beginning of the line

References :
* Kinnaird 2018 is not cited in the main text
* Mc fee 2019 is not cited in from the main text
* Mc Guirl is not cited in from the main text

In the official webpage :
* The terms “Recommended installation” and “general installation” seems not suitable. It seems more to be “install from source” and “install from pip or conda package-manager”, which should be the former proposition I guess. Also, in the part “install from source, you should mention that the user need first to git clone your repo.

Done for my first turn. Once the authors addresses these points, I’ll get a last turn to run the package exhaustively.
