### 2023-03-09 #5213 @ranzhengcode
@jbytecode Dear jbytecode, I have completely checked and tried repytah on my computer and was able to run all instances properly according to the author's instructions. As I am not an expert in music and data processing, I don't have a more pertinent advice for authors, but I think the work is still very interesting to accept.
