+++ JOSS-Paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: '`repytah`: An Open-Source Python Package That Builds Aligned Hierarchies for Sequential Data Streams'
+title: '`repytah`: An Open-Source Python Package for Building Aligned Hierarchies for Sequential Data'
 tags:
   - Python
   - Music Information Retrieval
@@ -58,7 +58,7 @@ We introduce `repytah`, a Python package that constructs the aligned hierarchies
 
 Broadly, Music Information Retrieval (MIR) seeks to capture information about music in pursuit of various music-related tasks, such as playlist recommendation, cover song identification, and beat detection. Content-based methods work directly on musical artifacts such as audio recordings or musical scores, while other approaches leverage information about musical artifacts such as metadata (like the artist or album name), tags (like genre), or listener surveys. Approaches to MIR tasks can also include a focus on repeated (or novel) elements. 
 
-Music-based data streams, like music, often have repeated elements that build on each other, creating structure-based hierarchies. The aligned hierarchies representation by Kinnaird `@Kinnaird_ah:2016` is a structure-based representation that places all possible structural-hierarchical decompositions of a data stream onto one common time axis. These aligned hierarchies when applied to music-based data streams can be used for visualization or for similarity tasks within MIR like the fingerprint task [@Kinnaird_ah]. The aligned hierarchies can also be post-processed to address additional tasks such as the cover song task [@Kinnaird_ash; @snl; @supp]. A drawback of Kinnaird’s code, however, is that the original code was written in MATLAB, which can only be used with a license and hence is not broadly accessible. 
+Music-based data streams, like music, often have repeated elements that build on each other, creating structure-based hierarchies. The aligned hierarchies representation by Kinnaird [@Kinnaird_ah:2016] is a structure-based representation that places all possible structural-hierarchical decompositions of a data stream onto one common time axis. These aligned hierarchies when applied to music-based data streams can be used for visualization or for similarity tasks within MIR like the fingerprint task [@Kinnaird_ah]. The aligned hierarchies can also be post-processed to address additional tasks such as the cover song task [@Kinnaird_ash; @snl; @supp]. A drawback of Kinnaird’s code, however, is that the original code was written in MATLAB, which can only be used with a license and hence is not broadly accessible. 
 
 There has been a long tradition of building Python packages for MIR research. Examples include the `AMEN` package [@amen], the `mir_eval` library [@mir_eval] , the `mirdata` library [@mirdata], and the more recent `libfmp` package [@libfmp]. As MIR has grown as a discipline, there has been an increased focus on reproducibility, accessibility, and open-source development [@reproducibility_mir, @mirdata]. As such, there have been several examples of code being translated from MATLAB to Python. The most notable example is `librosa`, a package that provides a number of powerful tools for MIR work [@librosa]. 
 
