+++ JOSS-Paper/paper.md
@@ -51,14 +51,14 @@ bibliography: joss.bib
 
 # Summary
 
-We introduce `repytah`, a Python package that constructs the aligned hierarchies representation that contains all possible structure-based hierarchical decompositions for a piece of sequential data aligned on a common time axis. In particular, this representation, introduced by Kinnaird [@Kinnaird_ah] with music-based data (like musical recordings or scores) at the primary motivation, is intended for sequential data where repetitions have particular meaning (such as a verse, chorus, motif, or theme). The `repytah` package builds these aligned hierarchies by first extracting repeated structures (of all meaningful lengths) from sequential data. This package is a Python translation of the original MATLAB code by Kinnaird [@Kinnaird_code] with additional documentation, and the code has been updated to leverage efficiencies in Python.
+We introduce `repytah`, a Python package that constructs the aligned hierarchies representation that contains all possible structure-based hierarchical decompositions for a piece of sequential data aligned on a common time axis. In particular, this representation, introduced by Kinnaird [@Kinnaird_ah] with music-based data (like musical recordings or scores) at the primary motivation, is intended for sequential data where repetitions have particular meaning (such as a verse, chorus, motif, or theme). The `repytah` package builds these aligned hierarchies by first extracting repeated structures (of all meaningful lengths) from sequential data. This package is a Python translation of the original MATLAB code by Kinnaird [-@Kinnaird_code] with additional documentation, and the code has been updated to leverage efficiencies in Python.
 
 
 # Statement of Need
 
 Broadly, Music Information Retrieval (MIR) seeks to capture information about music in pursuit of various music-related tasks, such as playlist recommendation, cover song identification, and beat detection. Content-based methods work directly on musical artifacts such as audio recordings or musical scores, while other approaches leverage information about musical artifacts such as metadata (like the artist or album name), tags (like genre), or listener surveys. Approaches to MIR tasks can also include a focus on repeated (or novel) elements. 
 
-Music-based data streams, like music, often have repeated elements that build on each other, creating structure-based hierarchies. The aligned hierarchies representation by Kinnaird [@Kinnaird_ah:2016] is a structure-based representation that places all possible structural-hierarchical decompositions of a data stream onto one common time axis. These aligned hierarchies when applied to music-based data streams can be used for visualization or for similarity tasks within MIR like the fingerprint task [@Kinnaird_ah]. The aligned hierarchies can also be post-processed to address additional tasks such as the cover song task [@Kinnaird_ash; @snl; @supp]. A drawback of Kinnaird’s code, however, is that the original code was written in MATLAB, which can only be used with a license and hence is not broadly accessible. 
+Music-based data streams, like music, often have repeated elements that build on each other, creating structure-based hierarchies. The aligned hierarchies representation by Kinnaird [-@Kinnaird_ah] is a structure-based representation that places all possible structural-hierarchical decompositions of a data stream onto one common time axis. These aligned hierarchies when applied to music-based data streams can be used for visualization or for similarity tasks within MIR like the fingerprint task [@Kinnaird_ah]. The aligned hierarchies can also be post-processed to address additional tasks such as the cover song task [@Kinnaird_ash; @snl; @supp]. A drawback of Kinnaird’s code, however, is that the original code was written in MATLAB, which can only be used with a license and hence is not broadly accessible. 
 
[…]
