### 2023-03-17 smith-tinkerlab/repytah#426 @kmkinnaird
Thank you for your detailed review. We have addressed your comments below:

> There is a claim of universality, stating that repytah can process any data stream, while examples are run on a single music file (a score). While I guess the claim could be valid, authors could modulate the claim (could be used… or something like that), or provide evidence with another dataset.

- Excellent point. In the newest version, we have modulated the claim in the summary.

> In addition, this could be minor, but the author state the tool process streams, and a user could take you to the word, and provide actual streams, i.e. data with no start and no end. I guess it is not suitable as repytah only processes finite matrices to manage a finite-size dissimilarity matrix. But as the term “stream” was used in the first paper from Kinnaird, I guess this data naming is accepted in the field.

- Thank you for bringing this up. We had not thought about the word "stream" from this perspective, but as you note in other disciplines "stream" implies either infinite or continuous generation. The text has been updated to move away from using the word stream.

> Also, the example file (in repytah/data) are CSV without mentioning what they are. Are they scores? Spectrum extracts? It is unclear. Please introduce a README in the data dir. I would like to understand what I processed precisely. This point could feed the first part of the question.

- A README has been added.

> L1: I am not good at English, and I guess Kinnaird master her language, but the word “That” in the title seems weird. It sounds like we found a package in the wilderness, plugged it in, and suddenly discovered that it behaves in that way. This point is a bit about “tastes and colors”, so it’s optional.

- A new title has been proposed.

> L4: I was confused with the extensive list of authors, but after an exhaustive search, it seems that they (quite) all contributed to the repo. This is nice, I admire the proficiency of the PI to make all them work together that way on a single object. However, I did not find out what Lillian Yushu Jiang was responsible for.

- This project was largely conducted by undergraduate students over a period of nearly four years. We have been fortunate to have several students involved through different funding sources and research opportunities, which is why the list of authors is so long. While many students contributed to the final produced code, others were involved in the project in other ways including working to understand the original thesis and how to translate that understanding into Python.

> L18 End of summary “and the code has been updated to leverage efficiencies in Python”, maybe better worded I guess, or gain a few words to get an insight of the efficiencies leveraged.

- We agree that this could be better worded. We were trying to draw a distinction between what MATLAB does well (eg. matrix operations) and what Python does well (eg. open-source, additional data structures, etc).

> L32 In the statement of need, maybe two or three more words to describe what is the “fingerprint task” and the “cover song task”, because here are your main “customers”.

- Additional definitions have been added.

> L33: the author should check the ref (Savard et al. 2020), which is not well spelled in the manuscript.

- Thank you for pointing this out. It appears that we had an error in our markdown file that produced this issue.

> L35 In the statement of need you state that the original code is MATLAB, thus implies licensing issue. But there are alternatives (octave). Did you try it out? Sometimes there are performance issues, or some functionalities of MATLAB are not integrated in octave. You should state this point, while the interest in having it available in python is still high.

- We did not consider octave in part because there is a pronounced preference for python in MIR and in other fields more broadly.

> L36 and L44 Speaking of a “tradition” of building MIR in Python, is maybe abusive as it is a 6 years sequence. Maybe we can speak of a trend, or an appeal from the community? Python growing interest?

- Text has been changed to say "recent trend" rather than "tradition."

> L40 : the author should check the ref (Bittner et al. 2019), which is not well spelled in the manuscript.

- We're not sure what you referring to. Could you clarify what you see as a misspelling?

> L45 “ for a given sequential data stream”, can you provide the dimensionality of the input space? I work on imaging, but I guess that repytah is not suitable for 2D+t, isn’t it ? And 3D+t ?

- Thank you for this question. It's clear that we haven't made plain that `repytah` can build the aligned hierarchies for any finite piece of sequential data that has been processed into a self-dissimilarity matrix (SDM). This was an intentional choice as an SDM can not be reverse-engineered back to the original signal. This means that one could use a sequence of images, but that you would have to first make the SDM for that sequence (choosing a notion of dissimilarity to do that).

> L49 : [...] by cross-referencing [...]. I don’t get it. What do the authors did? Did they run some comparison of the output of the Matlab package and the output of the python package, compute some comparison scores, or check exact matching? Are these data provided ? If not, it should be stated.

- The authors had a set of examples that they used to generate both formal unit tests as well as informal tests comparing the results from the MATLAB code to the python code. We did not generate comparison scores.

> L57: I don’t get the description of the module transform, you should make it clear.

- We have added additional language about this module and have shifted the order of how the modules are listed.

> L69 : what are “Chroma features” ? If repytah can be used for any data stream, you can understand that some of your potential users won’t understand this point. Can you add a few words about it? This point is linked to the request above for the data README.md

- We have added additional information about the features.

> L72: missing “funded by” at the beginning of the line

- Thank you for noting this. It has been corrected.

> References :
>   Kinnaird 2018 is not cited in the main text
>   Mc fee 2019 is not cited in from the main text
>   Mc Guirl is not cited in from the main text

- Thank you for pointing this out. It appears that we again had an error in our markdown file that produced this issue.

> In the official webpage :
>    The terms “Recommended installation” and “general installation” seems not suitable. It seems more to be “install from source” and “install from pip or conda package-manager”, which should be the former proposition I guess. Also, in the part “install from source, you should mention that the user need first to git clone your repo.

- We have made the recommended changes, but they will not update until the created branch is merged into `main`. We are keeping this branch separate for now during the review process.
