### 2023-03-17 smith-tinkerlab/repytah#426 @Rocsg
Thank you @kmkinnaird , I will check your points and run last tests on your package to fulfill my checklist early next week.
Wish you a nice weekend.

Romain
