+++ JOSS-Paper/joss.bib
@@ -7,7 +7,7 @@
       number = {63},
       pages = {3326},
       author = {M. Müller and F. Zalkow},
-      title = {libfmp: A Python Package for Fundamentals of Music Processing},
+      title = {libfmp: A {P}ython Package for Fundamentals of Music Processing},
       journal = {The Journal of Open Source Software}
 }
 
@@ -26,8 +26,8 @@
 
 @InProceedings{librosa,
       author = {B. McFee and C. Raffel and D. Liang and D. P. W. Ellis and M. McVicar and E. Battenberg and O. Nieto},
-      title = {librosa: Audio and Music Signal Analysis in Python},
-      booktitle = {In Proceedings of the 14th Python in Science Conference},
+      title = {librosa: Audio and Music Signal Analysis in {P}ython},
+      booktitle = {In Proceedings of the 14th {P}ython in Science Conference},
       pages = {18 - 24},
       year = {2015},
       editor = {K. Huff and J. Bergstra},
