### 2023-04-09 #5213 @jbytecode
@kmkinnaird - I've just sent a pull request that includes some minor changes on both the manuscript and the bibliography:

https://github.com/smith-tinkerlab/repytah/pull/428

Please review the changes. If you are agreed with them, please merge.

Thank you in advance.
