+++ JOSS-Paper/joss.bib
@@ -124,6 +124,7 @@
 @article{CS2006SEQ,
       author={M. Casey, and M. Slaney},
       title={The importance of sequences in music similarity},
+      doi = {10.1109/icassp.2006.1661198}
       journal=assp6,
       pages={V-5 - V-8},
       year=2006
