+++ JOSS-Paper/paper.md
@@ -51,7 +51,7 @@ bibliography: joss.bib
 
 # Summary
 
-We introduce `repytah`, a Python package that constructs the aligned hierarchies representation that contains all possible structure-based hierarchical decompositions for a finite length piece of sequential data aligned on a common time axis. In particular, this representation--introduced by Kinnaird [@Kinnaird_ah] with music-based data (like musical recordings or scores) as the primary motivation--is intended for sequential data where repetitions have particular meaning (such as a verse, chorus, motif, or theme). Although the original motivation for the aligned hierarchies representation was finding structure for music-based data streams, there is nothing inherent in the construction of these representations that limits `repytah` to only being used on sequential data that is music-based. 
+We introduce `repytah`, a Python package that constructs the aligned hierarchies representation that contains all possible structure-based hierarchical decompositions for a finite length piece of sequential data aligned on a common time axis. In particular, this representation--introduced by Kinnaird [-@Kinnaird_ah] with music-based data (like musical recordings or scores) as the primary motivation--is intended for sequential data where repetitions have particular meaning (such as a verse, chorus, motif, or theme). Although the original motivation for the aligned hierarchies representation was finding structure for music-based data streams, there is nothing inherent in the construction of these representations that limits `repytah` to only being used on sequential data that is music-based. 
 
 The `repytah` package builds these aligned hierarchies by first extracting repeated structures (of all meaningful lengths) from the self-dissimilarity matrix (SDM) for a piece of sequential data. Intentionally `repytah` uses the SDM as the starting point for constructing the aligned hierarchies, as an SDM cannot be reversed-engineered back to the original signal and allows for researchers to collaborate with signals that are protected either by copyright or under privacy considerations. This package is a Python translation of the original MATLAB code by Kinnaird [-@Kinnaird_code] with additional documentation, and the code has been updated to leverage efficiencies in Python. 
 
@@ -82,7 +82,7 @@ Each module has an associated Jupyter notebook file that summarizes the moduleâ€
 
 Additionally, the package includes `example.py` which runs a complete example building aligned hierarchies from a single example CSV file for the score of Chopin's Mazurka Op. 6, No. 1. This example CSV file stores the musical information about each beat of the Mazurka as Chroma Features (or the twelve Western tones: {C, C#, D, D#, etc}). These Chroma features are extracted from the associated files on the [kern Scores data base](https://kern.humdrum.org/cgi-bin/browse?l=users/craig/classical/chopin/mazurka) [@S2005].
 
-In `example.py`, the first steps include transforming the sequential data into the SDM by first shingling the features [@CS2007; @CS2006NEAR; CS2006SEQ] and then using the cosine-dissimilarity measure on the resulting shingles. Again, we note that while `example.py` does compute the SDM, the `repytah` package's main contribution is the construction of the aligned hierarchies from a given SDM. 
+In `example.py`, the first steps include transforming the sequential data into the SDM by first shingling the features [@CS2007; @CS2006NEAR; @CS2006SEQ] and then using the cosine-dissimilarity measure on the resulting shingles. Again, we note that while `example.py` does compute the SDM, the `repytah` package's main contribution is the construction of the aligned hierarchies from a given SDM. 
 
 
 # Acknowledgements
