### 2023-05-05 #5213 @kmkinnaird
Completing the post review checklist here:

- [X] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. --> **0.1.1**
- [X] Archive the release on Zenodo/figshare/etc and post the DOI here. --> **10.5281/zenodo.7898923**
