+++ repytah/version.py
@@ -6,7 +6,7 @@ import sys
 import importlib
 
 short_version = "0.1"
-version = "0.1.0dev"
+version = "0.1.2"
 
 
 def __get_mod_version(modname):
