+++ repytah/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "repytah" %}
-{% set version = "0.1.0" %}
+{% set version = "0.1.2" %}
 
 package:
   name: {{ name|lower }}
