### 2023-05-12 #5213 @kmkinnaird
@jbytecode I'm waiting for the conda-forge PR before I made the new release. It might take a few hours. Will check again tonight and post the new DOI here for the new release
