### 2023-05-13 #5213 @danielskatz
@kmkinnaird - I'm the track editor who will handle the final processing of your submission. I've proofread the draft, and the paper looks fine to me, but some references need work. Specifically, there are a few references that are of type article that should be of type inproceedings, such as Casey & Slaney 2006a, Casey & Slaney 2006b, Casey & Slaney 2007, and Sapp 2005. The conference name should also be written out in these cases, as it is in others already. Please fix these, then ping me here, and I'll continue the process of acceptance and publication.
