### 2023-05-15 616cc9d
updating bibliography with full conference names
