+++ JOSS-Paper/joss.bib
@@ -96,36 +96,36 @@
 }
 
 
-@article{S2005,
+@InProceedings{S2005,
       author={C.S. Sapp},
       title={Online Database of Scores in the Humdrum File Format},
-      journal=ismir06,
+      booktitle={In Proceedings of the 6th International Society for Music Information Retrieval (ISMIR) Conference},
       pages={664-665},
       year= 2005
 }
 
 
-@article{CS2007,
+@InProceedings{CS2007,
       author={M. Casey, and M. Slaney},
       title={Fast Recognition of Remixed Audio},
-      journal=assp7,
+      booktitle={IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP) Proceedings},
       pages={IV-1425 - IV-1428},
       year={2007},
 }
 
-@article{CS2006NEAR,
+@InProceedings{CS2006NEAR,
       author={M. Casey, and M. Slaney},
       title={Song Intersection by Approximate Nearest Neighbor Search},
-      journal=ismir07,
+      booktitle={In Proceedings of the 6th International Society for Music Information Retrieval (ISMIR) Conference},
       pages={144-149},
       year=2006,
 }
 
-@article{CS2006SEQ,
+@InProceedings{CS2006SEQ,
       author={M. Casey, and M. Slaney},
       title={The importance of sequences in music similarity},
       doi = {10.1109/icassp.2006.1661198},
-      journal=assp6,
+      booktitle={IEEE International Conference on Acoustics Speech and Signal Processing (ICASSP) Proceedings},
       pages={V-5 - V-8},
       year=2006
 }
