### 2023-05-15 3231554
updating bibliography with consistent naming for authors
