+++ JOSS-Paper/joss.bib
@@ -19,13 +19,13 @@
       volume = {36},
       number = {1},
       pages = {128-137},
-      author = {B. McFee and J. W. Kim and M. Cartwright and J. Salamon and R. M. Bittner and J. P. Bello},
+      author = {B. McFee and J.W. Kim and M. Cartwright and J. Salamon and R.M. Bittner and J. P. Bello},
       title = {Open source practices for music signal processing research},
       journal = {IEEE Signal Processing Magazine}
 }
 
 @InProceedings{librosa,
-      author = {B. McFee and C. Raffel and D. Liang and D. P. W. Ellis and M. McVicar and E. Battenberg and O. Nieto},
+      author = {B. McFee and C. Raffel and D. Liang and D.P.W. Ellis and M. McVicar and E. Battenberg and O. Nieto},
       title = {librosa: Audio and Music Signal Analysis in {P}ython},
       booktitle = {In Proceedings of the 14th {P}ython in Science Conference},
       pages = {18 - 24},
@@ -35,28 +35,28 @@
 }
 
 @InProceedings{mir_eval,
-      author = {C. Raffel and B. McFee and E. J. Humphrey and J. Salamon and O. Nieto and D. Liang and D. P. W. Ellis},
+      author = {C. Raffel and B. McFee and E.J. Humphrey and J. Salamon and O. Nieto and D. Liang and D.P.W. Ellis},
       title = {mir\_eval: A Transparent Implementation of Common MIR Metrics},
       booktitle = {In Proceedings of the 15th International Society for Music Information Retrieval (ISMIR) Conference},
       year = {2014}
 }
 
 @InProceedings{mirdata,
-      author = {R. M. Bittner and M. Fuentes and D. Rubinstein and A. Jansson and K. Choi and T. Kell},
+      author = {R.M. Bittner and M. Fuentes and D. Rubinstein and A. Jansson and K. Choi and T. Kell},
       title = {mirdata: Software for Reproducible Usage of Datasets},
       booktitle = {In Proceedings of the 20th International Society for Music Information Retrieval (ISMIR) Conference},
       year = {2019}
 }
 
 @InProceedings{Kinnaird_ah,
-      author = {K. M. Kinnaird},
+      author = {K.M. Kinnaird},
       title = {Aligned Hierarchies: A Multi-Scale Structure-Based Representation for Music-Based Data Streams},
       booktitle = {In Proceedings of the 17th International Society for Music Information Retrieval (ISMIR) Conference},
       year = {2016}
 }
 
 @InProceedings{Kinnaird_ash,
-      author = {K. M. Kinnaird},
+      author = {K.M. Kinnaird},
       title = {Aligned Sub-Hierarchies: A Structure-based Approach to the Cover Song Task},
       booktitle = {In Proceedings of the 19th International Society for Music Information Retrieval (ISMIR) Conference},
       year = {2018}
@@ -64,7 +64,7 @@
 
 
 @misc{Kinnaird_code,
-      author = {K. M. Kinnaird},
+      author = {K.M. Kinnaird},
       title = {Thesis Code},
       year = {2014},
       publisher = {GitHub},
@@ -73,14 +73,14 @@
 }
 
 @InProceedings{snl,
-      author = {M. R. McGuirl and K. M. Kinnaird and C. Savard and E. H. Bugbee},
+      author = {M.R. McGuirl and K.M. Kinnaird and C. Savard and E.H. Bugbee},
       title = {SE and SNL diagrams: Flexible data structures for MIR},
       booktitle = {In Proceedings of the 19th International Society for Music Information Retrieval (ISMIR) Conference},
       year = {2018}
 }
 
 @InProceedings{supp,
[…]
