### 2023-05-15 #5213 @kmkinnaird
@danielskatz - Thanks for your note and for introducing yourself.

I believe that I've fixed all the references to include the full conference names as well as consistent presentation for the author names. The current draft is in the `main` branch of the repo

One question: are the use of acronyms correct or should I remove them?
